refactor storage, and a commentary on claim permissions
Also removed weight from ClaimPermission. Idea behind weights was to
allow for addons to easily add in other permissions to the hierarchy -
however, since it's an enum that's probably not going to work out that
well. Most permission-related requests are very specific anyways, or are
in the form of individual flags, which don't fit in a hierarchy